main/requirements.txt
Flask==3.0.3
langchain==0.3.4
langchain_huggingface==0.1.0
langchain_community==0.3.3
chromadb==0.5.0
sentence-transformers==3.2.0
unstructured==0.14.4
google-generativeai==0.8.3
markdown==3.7
python-dotenv==1.0.1
sentencepiece==0.2.0
groq==0.11.0